+++ src/Zeus/Customizer/Exception/CustomizerException.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace GetOlympus\Zeus\Customizer\Exception;
+
+use GetOlympus\Zeus\Base\Exception\BaseException;
+
+/**
+ * Customizer exception.
+ *
+ * @package    OlympusZeusCore
+ * @subpackage Customizer\Exception
+ * @author     Achraf Chouk <achrafchouk@gmail.com>
+ * @since      2.0.1
+ *
+ */
+
+class CustomizerException extends BaseException
+{
+    //
+}